refact: hook 폴더로 이동 및 any 타입 구체적으로 명시
